edited SQl-query in RoomRepository

+++ src/main/java/com/lap/lapproject/repos/BookingRepository.java
@@ -9,6 +9,6 @@ import java.util.ArrayList;
 public interface BookingRepository {
 
     ArrayList<Booking> readAll() throws SQLException;
-    void addBooking(long roomID, long userID, long trainerID, long courseID,
+    void addBooking(int roomID, int userID, int trainerID, int courseID,
                     String recurrenceRule, LocalDateTime localDateTimeStart, LocalDateTime localDateTimeEnd) throws SQLException;
 }
